Seperated move history logic into a hook
